Refine solutions.

+++ nth-prime/src/iter_primes.rs
@@ -6,8 +6,8 @@ use std::ops::RangeFrom;
 
 // Priority queue–based algorithm from https://doi.org/10.1017/S0956796808007004
 pub struct IterPrimes {
-    compositeses: BinaryHeap<Reverse<OrdByItem<StepBy<RangeFrom<u64>>>>>,
-    naturals: RangeFrom<u64>,
+    compositeses: BinaryHeap<Reverse<OrdByItem<StepBy<RangeFrom<u32>>>>>,
+    naturals: RangeFrom<u32>,
 }
 
 impl IterPrimes {
@@ -20,16 +20,18 @@ impl IterPrimes {
 }
 
 impl Iterator for IterPrimes {
-    type Item = u64;
+    type Item = u32;
 
-    fn next(&mut self) -> Option<u64> {
+    fn next(&mut self) -> Option<u32> {
         let compositeses = &mut self.compositeses;
 
         self.naturals.find(|&n| {
             let is_prime = compositeses.peek().map(|c| c.0.item) != Some(Some(n));
 
             if is_prime {
-                compositeses.push(Reverse(OrdByItem::new((n.pow(2)..).step_by(n as usize))));
+                compositeses.push(Reverse(OrdByItem::new(
+                    (u64::from(n).pow(2) as u32..).step_by(n as usize),
+                )));
             } else {
                 while let Some(mut composites) = compositeses.peek_mut() {
                     // Pending eRFC 2497
